Converting account invoices block
