--- blocks/i18n.js
@@ -1,1 +1,0 @@
-wp.i18n.setLocaleData( { '': {} }, 'paid-memberships-pro' );
