+++ blocks/src/account-invoices-section/render.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @see https://github.com/WordPress/gutenberg/blob/trunk/docs/reference-guides/block-api/block-metadata.md#render
+ */
+$title = isset( $attributes['title'] ) ? $attributes['title'] : null;
+$output = pmpro_shortcode_account( array( 'sections' => 'invoices', 'title' => $title ) );
+
+?>
+<p <?php echo get_block_wrapper_attributes(); ?>>
+	<?php echo $output; ?>
+</p>
