+++ blocks/src/account-invoices-section/style.scss
@@ -1,0 +1,12 @@
+/**
+ * The following styles get applied both on the front of your site
+ * and in the editor.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
+
+.wp-block-create-block-account-invoices-section {
+	background-color: #21759b;
+	color: #fff;
+	padding: 2px;
+}
